@@ -2,11 +2,11 @@ import React from "react";
 
 import "./arrow.scss";
 
-interface ArrowProps extends React.HTMLAttributes<HTMLDivElement> {
+interface Props extends React.HTMLAttributes<HTMLDivElement> {
   direction: string;
 }
 
-function Arrow(props: ArrowProps) {
+function Arrow(props: Props) {
   return <div className={`arrow arrow_${props.direction}`}></div>;
 }
 
