@@ -4,17 +4,17 @@ import { NavLink } from "react-router-dom";
 import "./btn.scss";
 
 interface Props extends React.HTMLAttributes<HTMLLinkElement> {
-  href: string;
-  text: string;
-  handleClick?: (e: any) => void;
+  to: string;
+  children: React.ReactNode;
+  className?: string;
 }
 
 function Btn(props: Props) {
-  const { className, href, text } = props;
+  const { className = "", to, children } = props;
 
   return (
-    <NavLink to={href} className={className} onClick={props.handleClick}>
-      {text}
+    <NavLink to={to} className={`btn btn_theme_empty ${className}`}>
+      {children}
     </NavLink>
   );
 }
