@@ -1,38 +1,38 @@
-import React, { Component } from "react";
+import React from "react";
 
-import { ReleaseMetadata } from "../../types";
 import { Release } from "../release/release";
 
 import "./content-grid.scss";
 
-interface Props extends React.HTMLAttributes<HTMLDivElement> {
+interface Props {
   releases: ReleaseMetadata[];
-  handleDeleteReleaseBtnClick: (releaseId: number) => void;
+  playingTrackId?: number;
+
+  togglePlay: (metadata: TrackExtendedMetadata) => void;
+
+  className?: string;
 }
 
-class ContentGrid extends Component<Props> {
-  constructor(props: Props) {
-    super(props);
-  }
+//
 
-  render() {
-    return (
-      <main className="content-grid">
-        {...this.props.releases.map((release) => {
-          return (
-            <Release
-              metadata={release}
-              releaseId={release.id}
-              key={release.id.toString()}
-              handleDeleteReleaseBtnClick={
-                this.props.handleDeleteReleaseBtnClick
-              }
-            />
-          );
-        })}
-      </main>
-    );
-  }
+export function ContentGrid(props: Props): JSX.Element {
+  const { className = "" } = props;
+
+  console.log(props.playingTrackId);
+
+  return (
+    <div className={`content-grid ${className}`}>
+      {...props.releases.map((release) => {
+        return (
+          <Release
+            togglePlay={props.togglePlay}
+            playingTrackId={props.playingTrackId}
+            metadata={release}
+            releaseId={release.id}
+            key={release.id.toString()}
+          />
+        );
+      })}
+    </div>
+  );
 }
-
-export { ContentGrid };
