@@ -1,13 +1,13 @@
-import React from "react";
+import React, { PropsWithChildren } from "react";
 
 import "./error.scss";
 
-interface ErrorProps extends React.HtmlHTMLAttributes<HTMLDivElement> {}
-
-function Error(props: ErrorProps) {
-  const { className = "error", children } = props;
-
-  return <div className={className}>{children}</div>;
+interface Props {
+  className?: string;
 }
 
-export { Error };
+export function Error(props: PropsWithChildren<Props>): React.ReactNode {
+  const { className = "", children } = props;
+
+  return <span className={`error ${className}`}>{children}</span>;
+}
