@@ -7,12 +7,12 @@ interface MusicboxLogoProps extends React.HTMLAttributes<SVGAElement> {
 }
 
 function MusicboxLogo(props: MusicboxLogoProps) {
-  const { className = "app-logo", fill = "black", height = "1.5rem" } = props;
+  const { className, fill = "white", height = "1.5rem" } = props;
   const SVGpath =
     "m 42.005218,0 c -0.93917,0 -1.740234,0.67676207 -1.740234,1.546875 0,0.870113 0.801064,1.546875 1.740234,1.546875 0.93917,0 1.740235,-0.676762 1.740235,-1.546875 C 43.745453,0.67676207 42.944388,0 42.005218,0 Z m 8.613281,0.19335938 V 10.34375 h 2.998047 V 9.6542969 c 0.359095,0.4281511 1.007819,0.9804691 2.167969,0.9804691 1.546868,0 3.052734,-1.3125117 3.052734,-3.6328129 0,-2.4722256 -1.616357,-3.6191406 -3.052734,-3.6191406 -1.173962,0 -1.809851,0.5947279 -2.072266,0.8847656 V 0.19335938 Z M 1.9316406,0.3984375 V 1.2636719 H 9.5742188 V 0.3984375 Z m -0.75,1.5449219 v 0.00195 0.8652344 H 10.445312 V 1.9433594 Z M 18.548187,3.3828125 c -0.952981,0 -1.698524,0.4429438 -2.154297,1.0644531 V 3.6738281 H 13.493499 V 10.34375 H 16.39389 V 6.9609375 c 0,-0.8010564 0.331335,-1.2578125 0.787109,-1.2578125 0.428151,0 0.566407,0.3462637 0.566407,0.8710938 V 10.34375 h 2.90039 V 6.9609375 c 0,-0.8010564 0.331337,-1.2578125 0.78711,-1.2578125 0.428151,0 0.566406,0.3462637 0.566406,0.8710938 V 10.34375 h 2.900391 V 6.4492188 c 0,-1.0082261 -0.16519,-3.0664063 -2.375,-3.0664063 -0.994414,0 -1.629328,0.3600759 -2.167969,1.0644531 -0.469585,-0.828679 -1.133792,-1.0644531 -1.810547,-1.0644531 z m 18.460937,0 c -2.113131,0 -3.039062,1.1199838 -3.039062,2.390625 0,1.2982637 0.759772,1.7946439 2.113281,2.015625 0.621509,0.1104905 0.71875,0.2767869 0.71875,0.5253906 0,0.041434 -0.01451,0.3867188 -0.705078,0.3867188 -0.566264,0 -0.966661,-0.2075909 -1.560547,-0.5390625 l -0.773437,1.7675781 c 0.814868,0.5110185 1.533069,0.7050785 2.707031,0.7050785 3.135169,0 3.232422,-1.934573 3.232422,-2.4179691 0,-0.8148677 -0.26354,-1.7806932 -1.824219,-2.0292969 -0.511018,-0.082868 -1.103516,-0.096821 -1.103516,-0.5664062 0,-0.1104907 0.05552,-0.4980469 0.746094,-0.4980469 0.372906,0 0.856448,0.1396498 1.339844,0.4296875 l 0.74414,-1.546875 C 38.513734,3.3981615 37.243917,3.3828125 37.009124,3.3828125 Z m 10.863282,0 c -2.251245,0 -3.63086,1.6441224 -3.63086,3.6191406 0,2.2098107 1.586785,3.6328129 3.63086,3.6328129 1.008226,0 1.726843,-0.304828 2.085937,-0.484375 V 7.5957031 c -0.593886,0.4419622 -0.925505,0.5253907 -1.339844,0.5253907 -0.690566,0 -1.380859,-0.3168284 -1.380859,-1.0488282 0,-0.8148677 0.636164,-1.1757812 1.271484,-1.1757812 0.607698,0 1.131559,0.3332881 1.449219,0.5957031 v -2.625 C 49.198721,3.5357159 48.604405,3.3828125 47.872406,3.3828125 Z m 15.326171,0 c -2.831319,0 -4.033203,1.7269903 -4.033203,3.6191406 0,1.9059617 1.215695,3.6328129 4.033203,3.6328129 2.81751,0 4.033204,-1.7268512 4.033204,-3.6328129 0,-1.8921503 -1.201884,-3.6191406 -4.033204,-3.6191406 z M 0.56835938,3.515625 V 4.4746094 H 11.054688 V 3.515625 Z M 25.755218,3.6738281 V 7.265625 c 0,0.8010564 0.07005,3.369141 3.771485,3.369141 3.701432,0 3.769531,-2.5680846 3.769531,-3.369141 V 3.6738281 h -3.09375 v 3.4941407 c 0,0.3314716 -0.01284,0.953125 -0.675781,0.953125 -0.662943,0 -0.677735,-0.6216534 -0.677735,-0.953125 V 3.6738281 Z m 14.703125,0 V 10.34375 h 3.09375 V 3.6738281 h -1.546875 z m 26.355487,0 2.416015,3.1347657 -2.707031,3.5351562 c 0,0 2.188137,0 3.191406,0 l 1.160156,-1.890625 1.160157,1.890625 h 3.189453 L 72.516954,6.8085938 74.934923,3.6738281 H 71.647814 L 70.874376,5.2617188 70.100939,3.6738281 Z M 0,5.1855469 V 10.347656 H 11.626953 V 5.1855469 Z M 54.679046,5.703125 c 0.676754,0 1.160156,0.5253944 1.160157,1.2988281 0,0.8010564 -0.483403,1.3125 -1.160157,1.3125 -0.676755,0 -1.160156,-0.5114436 -1.160156,-1.3125 0,-0.7734337 0.483401,-1.2988281 1.160156,-1.2988281 z m 8.519531,0.1933594 c 0.801058,0 1.035157,0.6911292 1.035157,1.1054687 0,0.4281508 -0.234099,1.1191406 -1.035157,1.1191407 -0.801056,0 -1.037109,-0.6909899 -1.037109,-1.1191407 0,-0.4143395 0.236053,-1.1054687 1.037109,-1.1054687 z";
 
   return (
-    <NavLink exact to="/" className={className}>
+    <NavLink exact to="/" className={`musicbox-logo ${className}`}>
       <svg
         role="img"
         aria-label="Logo"
