@@ -1,60 +1,43 @@
-import React, { Component, Fragment } from "react";
+import React, { Fragment } from "react";
 
-interface SelectSortProps extends React.HTMLAttributes<HTMLDivElement> {
+interface Props extends React.HTMLAttributes<HTMLDivElement> {
   value: string;
-  onSelectSortChange: (controlName: string, value: string) => void;
-
-  gridLayoutActive: boolean;
-  listLayoutActive: boolean;
-}
-interface SelectSortState {}
-
-class SelectSort extends Component<SelectSortProps, SelectSortState> {
-  constructor(props: SelectSortProps) {
-    super(props);
-
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange(e: React.ChangeEvent<HTMLSelectElement>) {
-    this.props.onSelectSortChange(e.target.name, e.target.value);
-  }
-
-  render() {
-    const { className = "select-sort" } = this.props;
-
-    return (
-      <div className={className}>
-        Sort{" "}
-        <select
-          name="sort"
-          className="select-sort__box"
-          onChange={this.handleChange}
-          value={this.props.value}
-        >
-          {this.props.gridLayoutActive ? (
-            <Fragment>
-              <option value="artist,asc">Artist, A-Z</option>
-              <option value="artist,desc">Artist, Z-A</option>
-              <option value="title,asc">Release Title, A-Z</option>
-              <option value="title,desc">Release Title, Z-A</option>
-              <option value="year,asc">Year, 0-9</option>
-              <option value="year,desc">Year, 9-0</option>
-            </Fragment>
-          ) : (
-            <Fragment>
-              <option value="track-artist,asc">Artist, A-Z</option>
-              <option value="track-artist,desc">Artist, Z-A</option>
-              <option value="track-title,asc">Title, A-Z</option>
-              <option value="track-title,desc">Title, Z-A</option>
-              <option value="year,asc">Year, 0-9</option>
-              <option value="year,desc">Year, 9-0</option>
-            </Fragment>
-          )}
-        </select>
-      </div>
-    );
-  }
+  onSelectSort: (selected: string) => void;
+  layout: Layout;
 }
 
-export { SelectSort };
+export function SelectSort(props: Props) {
+  const { className = "" } = props;
+
+  return (
+    <div className={`select-sort ${className}`}>
+      Sort{" "}
+      <select
+        name="sort"
+        className="select-sort__box"
+        onChange={(e) => props.onSelectSort(e.target.value)}
+        value={props.value}
+      >
+        {props.layout === "grid" ? (
+          <Fragment>
+            <option value="artist,asc">Artist, A-Z</option>
+            <option value="artist,desc">Artist, Z-A</option>
+            <option value="title,asc">Release Title, A-Z</option>
+            <option value="title,desc">Release Title, Z-A</option>
+            <option value="year,asc">Year, 0-9</option>
+            <option value="year,desc">Year, 9-0</option>
+          </Fragment>
+        ) : (
+          <Fragment>
+            <option value="track-artist,asc">Artist, A-Z</option>
+            <option value="track-artist,desc">Artist, Z-A</option>
+            <option value="track-title,asc">Title, A-Z</option>
+            <option value="track-title,desc">Title, Z-A</option>
+            <option value="year,asc">Year, 0-9</option>
+            <option value="year,desc">Year, 9-0</option>
+          </Fragment>
+        )}
+      </select>
+    </div>
+  );
+}
