@@ -1,4 +1,5 @@
-declare module "*.svg" {
-  const content: any;
-  export default content;
-}
+declare module "*.jpg";
+declare module "*.png";
+declare module "*.jpeg";
+declare module "*.gif";
+declare module "*.svg";
